sjbackend/app/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
mysql-connector-python==8.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
redis==5.0.1
alembic==1.12.1
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
email-validator==2.1.0
bcrypt==4.0.1
python-json-logger==2.0.7EOF

# Install all dependencies
#pip install -r requirements.txt